+++ src/Resources/contao/dca/tl_content.php
@@ -1,4 +1,4 @@
-<?php if (!defined('TL_ROOT')) die('You cannot access this file directly!');
+<?php
 
 /**
  * Contao Open Source CMS
@@ -14,7 +14,9 @@
 /**
  * Palettes
  */
-$GLOBALS['TL_DCA']['tl_content']['palettes']['domgrabber'] = '{type_legend},type,headline;{domgrabber_legend},domgrabber_url,domgrabber_element,domgrabber_css;{protected_legend:hide},protected;{expert_legend:hide},guest,cssID;{invisible_legend:hide},invisible,start,stop';
+$GLOBALS['TL_DCA']['tl_content']['palettes']['__selector__'][] = 'includeCache';
+$GLOBALS['TL_DCA']['tl_content']['palettes']['domgrabber'] = '{type_legend},type,headline;{domgrabber_legend},domgrabber_url,domgrabber_element,domgrabber_css;{cache_legend:hide},includeCache;{protected_legend:hide},protected;{expert_legend:hide},guest,cssID;{invisible_legend:hide},invisible,start,stop';
+$GLOBALS['TL_DCA']['tl_content']['subpalettes']['includeCache'] = 'cache';
 
 /**
  * Fields
@@ -58,4 +60,32 @@ $GLOBALS['TL_DCA']['tl_content']['fields']['domgrabber_css'] = array
 		'rte'                => 'ace|css'
 	),
 	'sql'                    => "mediumtext NULL"
+);
+
+$GLOBALS['TL_DCA']['tl_content']['fields']['includeCache'] = array
+(
+	'label'                  => &$GLOBALS['TL_LANG']['tl_content']['includeCache'],
+	'inputType'               => 'checkbox',
+	'eval'                    => array
+	(
+		'submitOnChange'      => true
+	),
+	'sql'                     => array
+	(
+		'type'                => 'boolean',
+		'default'             => false
+	)
+);
+
+$GLOBALS['TL_DCA']['tl_content']['fields']['cache'] = array
+(
+	'label'                  => &$GLOBALS['TL_LANG']['tl_content']['cache'],
+	'search'                  => true,
+	'inputType'               => 'select',
+	'options'                 => &$GLOBALS['TL_LANG']['tl_content']['cache_optionen'],
+	'eval'                    => array
+	(
+		'tl_class'            => 'w50 clr'
+	),
+	'sql'                     => "int(10) unsigned NOT NULL default 0"
 );